+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-[Unreleased]: https://github.com/datenlord/aligned-utils/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/datenlord/aligned-utils/compare/v1.0.0...HEAD
+
+## [1.0.0] - 2021-01-12
+
+[1.0.0]: https://github.com/datenlord/aligned-utils/compare/v0.2.0...v1.0.0
 
 ## Added